Problem: Too much going on during eval
Because of how default.nix is set up, any use of buildRacket at all involves bootstrapping racket2nix before you can get a derivation or even a function.

Solution: Make racket2nix from a static nix expression.

We check that the expression is up to date in release, just like with
racket-packages, and we bootstrap racket2nix in release.