Problem: No cycle detection in Nix
Cycle detection in racket2nix is too early. If we want to have
overrides that manipulate dependencies, the cycle detection will be
based on outdated information and we may introduce infinite recursion.

Solution: Implement cycle detection in Nix.

We are using Tarjan's algorithm with minor changes:

 - The original detects cycles in the entire set, by randomly
   searching from untouched nodes until all nodes have been touched.
   We lazily detect cycles in the subtree that interests us.
 - The mathematical invariants used in Tarjan's proof require nodes to
   record lowlink as a found transitive parent's index rather than its
   lowlink. The code is slightly neater if you always use the
   lowlink of nodes you find, and the result is the same.

We are not yet using the cycle detection, as racket2nix provides
derivations rather than strings as thinRacketBuildInputs.